+++ OrderApi/Repository/OrderRepository.cs
@@ -1,7 +1,7 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using Library.Extensions;
+using Microsoft.EntityFrameworkCore;
 using OrderApi.Data;
 using OrderApi.Models;
-using OrderApi.Models.Extensions;
 
 namespace OrderApi.Repository
 {
@@ -17,9 +17,16 @@ namespace OrderApi.Repository
             _message = string.Empty;
         }
 
-        public async Task<PaginatedResult<Order>> GetAllPaginatedAsync(int pageNumber, int pageSize)
+        public async Task<PaginatedResult<Order>> GetAllPaginatedAsync(int pageNumber, int pageSize, string searchTerm, Filter? filter)
         {
-            IEnumerable<Order> orders = await _context.Orders.ToListAsync();
+            IEnumerable<Order> orders;
+            if (!string.IsNullOrWhiteSpace(searchTerm))
+                orders = await SearchEntitiesAsync(searchTerm);
+            else
+                orders = await _context.Orders.AsNoTracking().ToListAsync();
+
+            if (orders.Any() && filter != null)
+                orders = await FilterEntitiesAsync(orders, filter);
 
             var totalOrders = await Task.FromResult(orders.Count());
 
@@ -43,9 +50,46 @@ namespace OrderApi.Repository
             };
         }
 
+        public async Task<IEnumerable<Order>> SearchEntitiesAsync(string searchTerm)
+        {
+            var orders = await _context.Orders
+                .AsNoTracking()
+                .Where(o => o.Address.Contains(searchTerm) || 
+                            o.Region.Contains(searchTerm) ||
+                            o.City.Contains(searchTerm))
+                .ToListAsync();
+            return orders;
+        }
+
+        public async Task<IEnumerable<Order>> FilterEntitiesAsync(IEnumerable<Order> orders,Filter filter)
+        {
+            if (filter.OrderDateStart.HasValue)
+                orders = orders.Where(o => o.OrderDate >= filter.OrderDateStart.Value);
+
+            if (filter.OrderDateEnd.HasValue)
+                orders = orders.Where(o => o.OrderDate <= filter.OrderDateEnd.Value);
+
+            if (filter.DeliveryDateStart.HasValue)
+                orders = orders.Where(o => o.DeliveryDate >= filter.DeliveryDateStart);
+
+            if (filter.DeliveryDateEnd.HasValue)
+                orders = orders.Where(o => o.DeliveryDate <= filter.DeliveryDateEnd);
+
+            if (filter.Status.HasValue)
+                orders = orders.Where(o => o.Status == filter.Status);
+
+            if (filter.DeliveryId.HasValue)
+                orders = orders.Where(o => o.DeliveryTypeId == filter.DeliveryId);
+
+            return await Task.FromResult(orders.ToList());
+        }
+
+
         public async Task<Order?> GetByIdAsync(Guid id)
         {
-            var order = await _context.Orders.FirstOrDefaultAsync(o => o.OrderId == id);
+            var order = await _context.Orders
+                .AsNoTracking()
+                .FirstOrDefaultAsync(o => o.OrderId == id);
             if (order == null)
             {
                 _message = $"Order with Id [{id}] not found.";
@@ -106,21 +150,15 @@ namespace OrderApi.Repository
             _logger.LogInformation($"Order with Id[{order.OrderId}] updated succesfully.");
         }
 
-        public async Task DeleteAsync(Order order)
+        public async Task DeleteAsync(Guid id)
         {
-            try
-            {
-                _context.Orders.Remove(order);
-                await _context.SaveChangesAsync();
-            }
-            catch (Exception ex)
-            {
-                _message = $"Deletion of Order with id [{order.OrderId}] has failed.";
-                _logger.LogError(_message);
-                throw new ArgumentException(_message, ex);
-            }
+            var order = await _context.Orders.FindAsync(id);
 
-            _logger.LogInformation($"Order with Id [{order.OrderId}] deleted succesfully.");
+            if (order == null)
+                throw new KeyNotFoundException();
+
+            _context.Orders.Remove(order);
+            await _context.SaveChangesAsync();
         }
     }
